move NodeBlob to Core
