+++ Packages/essential/Runtime/Core/NodeBlob.cs.meta
