+++ Packages/essential/Runtime/Entities/BehaviorTreeComponents.cs
@@ -2,6 +2,7 @@
 using Unity.Collections.LowLevel.Unsafe;
 using Unity.Entities;
 using Unity.Jobs;
+using EntitiesBT.Core;
 
 namespace EntitiesBT.Entities
 {
