+++ Packages/essential/Runtime/Entities/BehaviorTreeLifeCycleSystem.cs
@@ -1,4 +1,5 @@
 ﻿using System.Reflection;
+using EntitiesBT.Core;
 using Unity.Entities;
 
 namespace EntitiesBT.Entities
